Pixel scaling functions and renderer color getter function added and fixed Color's type

+++ src/renderer/color.h
@@ -8,7 +8,7 @@
 #pragma once
 
 typedef struct {
-    uint r, g, b, a;
+    uint8_t r, g, b, a;
 } Deccan_Color;
 
 typedef struct _priv_ColorList {
